Use Karma test runner; remove Grunt dependency
Remove all the Grunt dependencies and switch to using the Karma test
runner.

Setup the Karma configuration file to load all the tests and library
dependencies. It's configured to use the RequireJS and Jasmine adapters.
Unit tests will be run and watched in Chrome and Firefox when
`karma start` is run during develop. CI tests are run once in PhantomJS.

The `test-main.js` file is needed to configure RequireJS for Karma, and
filter out the tests in the list of files provided to Karma.